requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.6.0
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-mail==1.4.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
